Avoid call to getCanonicalPath()
Profiling suggests that Base100_Java8_config was spending almost 1% of
its time in getCanonicalPath() in AgpVersionChecker's compareVersions()
method. We don't need to call getCanonicalPath() on every call of
compareVersions(), only when we throw an exception, so this change
modifes the code to only call getCanonicalPath() in case of an
exception.

Bug: NA
Test: existing
Change-Id: Ib02aa27af18a7aa565baa758ed38a063a16f554f