Extend C++ build macro definition table to ndk-build
This table was originally for CMake cohabitation but, with
variant folding, it is extended to also work with ndk-build.
This CL does a few things to help make this table more
manageable:
- The ABI "lookup" functionality is now table-driven rather
  than ad hoc. This is to avoid mistakes and inconsistencies.
- A naming scheme is now enforced. For example, macros that
  apply to a value in CxxVariantModel must now start with
  NDK_VARIANT_*.
There should be no behavior change. At this point ndk-build-
based projects don't use macros yet. That will come in a
future CL.

Bug: 170485962
Test: Refactoring

Change-Id: I3aff8c4985a84ec0badd876dd08f2ab45bed2992