Add matrix test for AndroidX/Jetifier property combinations
Depending on the values of the `android.useAndroidX` and
`android.enableJetifier` properties and the contents of the dependency
graph, we will produce different types of errors/warnings.

This commit adds a matrix test to ensure we cover all possible cases
during the transition to removal of the Jetifier feature (and also a
plan to enable `useAndroidX` by default).

Currently, this test covers the following cases:
  + AndroidX=false, Jetifier=false: Covered
  + AndroidX=false, Jetifier=true: Covered
  + AndroidX=true, Jetifier=false: To add in a later commit
  + AndroidX=true, Jetifier=true: To add in a later commit

Bug: 167700246
Test: New AndroidXJetifierMatrixTest
Change-Id: I25710552ca3108b6e2cd69c28fdea45d1f2e16b7