Force PackageAndroidArtifact to run non-incrementally for release builds
This change is a modified version of reverted change
I1038585ce92e28101c2cf4291c783a36cbb44f63, which was reverted because
it had a larger-than-expected impact on the performance of debug builds.

Force PackageAndroidArtifact to run non-incrementally unless building a
debuggable variant. Running the task non-incrementally causes the APKs
to be deterministic and doesn't allow any virtual entries in the APKs.

This change will worsen the performance of incremental builds (except
debug builds).

This change was implemented by adding all incremental file inputs again
as non-incremental file inputs when the task is to be run
non-incrementally. This is a workaround until Gradle adds support for
this use case (https://github.com/gradle/gradle/issues/16976)

Bug: 172762362
Test: DeterministicApkTest, TaskMethodModifiersAndAnnotationsTest
Change-Id: I11c67230512f761e7157d1b3615e92a7f21a8957