Lint unit test normalization
This test updates the test result formatting code to share
a single normalization method, which replaces $ characters
and strips directory roots from the test output (which is
now optional from the test task).

When lint prints a text report in unit tests, it specically
annotates secondary locations with their corresponding message.
Unfortunately there was a typo in this message. In this CL
the typo is fixed, and more importantly the test differ is
made aware of this such that it doesn't fail all the tests
that already have the typo.

Bug: None
Test: Existing
Change-Id: I9a7f75182de25b7113ba771db82fb23499fdfa1b