Move Incident from lint-cli to lint-api
In the past, lint detectors would report issues by passing in
4 arguments (message, location, etc). Over time a few more were
added, such as quickfix descriptors.

Over in the CLI there was a wrapper object for these, used to
pass around and generate multiple lint reports such as text and
XML.

This CL moves that class into the lint API. Not only will this
let us have a single report(Incident) method where in the future
we can add additional metadata, but in an upcoming CL this will
also be used to pass previously reported incidents back to
detectors for additional filtering.

Test: Existing
Bug: None
Change-Id: I63fb6533f34ead8006c1d1a260e850e6e4c0c065