Don't skip DeviceProviderInstrumentTestTask
DeviceProviderInstrumentTestTask is used to be skipped
when there is no java source files. This assumption is
not true because there might be a compiler generated
code. This patch changes testsFound() method to check
existence of the test APK. Although this is suboptimal
solution, it is safe and better than tests doesn't run
at all.

Test: KotlinAppConnectedTest
Bug: 173770818
Change-Id: I9af77a660503a96dc3893ca18b3be9f3c14c2f82