Add lint map abstraction
For an upcoming change, lint needs to offer a way to store
client properties for detectors, persistently. This CL
adds a new map class intended to be used for these purposes.

It also adds this map as an optional member of the Incident
class, such that you can store data with an incident, as well
as some other improvements to Incident such as convenience
constructors and improved documentation.

Test: Existing
Bug: None

Change-Id: Id182add6261fcb9ec2b72e7096d2914638afd430