Remove lint support for really old minSdkVersions (<15)
Studio (and AndroidX etc) have required at least API level 15
as minSdkVersion for many years now.

There were a number of places where lint was checking for
obsolete minSdkVersions to add or remove warnings. This
CL removes those code paths. (This simplifies an upcoming
change which deals with minSdkVersions.)

It also fully deletes a couple of lint checks that are
irrelevant at this point (FieldGetter, ViewTag).

Test: Existing (and updated, removing tests for old API levels)
Bug: None
Change-Id: Ie4d4ded4fbea33b3e4000c74a935fd526d4ef52d