Remove scheduling events from Swapper
Filter out scheduling events from Swapper thread (it does not have a
process), since it's used at system level to represent a dummy
process/thread that can be swapped by the kernel into some real
workload.

We filter here instead of inside Studio, because it produces a huge
number of events (any time a core is idle with no thread scheduled on
it) and bloats the proto response.

Bug: 168840323
Test: scheduling_request_handler_test.cc + visual
Change-Id: I99c18ae8692ee21bf6de1620572b0645c89fad4c