Fix missing test coverage file with clearPackageData option
There were two problems.
1) AGP retrieves the test coverage file path from
"-e coverageFile" params where it should be
"-e coverageFilePath" with the Android test orchestrator.

2) AGP deletes and recreate a directory when you set a
directory in "-e coverageFilePath" param. The task executed
"rm -r dir", the rm command outputs a confirmation
message and the task waits for the standard input forever.
This caused the hang in debugAndroidTest task. This patch
adds "-f" option to "rm" command. It also adds "-p" option
to "mkdir" command to ensure the parent directories are
created if missing.

With this fix, you can enable test coverage with clearPackageData
option.

android {
    defaultConfig {
        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
        testInstrumentationRunnerArguments clearPackageData: 'true'
        testInstrumentationRunnerArguments coverageFilePath: '/storage/emulated/0/coverage'
    }
    buildTypes {
        debug {
            testCoverageEnabled = true
        }
    }
    testOptions {
        execution 'ANDROIDX_TEST_ORCHESTRATOR'
    }
}

Although there are two limitations:

1) You must use non application private directory as coverageFilePath
in order for the coverage data not to be deleted after "pm clear"
command by the clearPackageData option. This also means that
your application has to declare android.permission.WRITE_EXTERNAL_STORAGE
permission in the AndroidManifest.xml file.

2) You have to grant WRITE_EXTERNAL_STORAGE at runtime for each test case.
The easiest way to do this is using the test rule,
GrantPermissionRule.grant(Manifest.permission.WRITE_EXTERNAL_STORAGE).
This is required because after the Android test orchestrator calls
"pm clear" command, those granted permissions are reset. In order for
tested application to be able to write the coverage data onto the
external file storage, you have to re-grant the permission.

Test: SimpleTestRunnableTest and manually (I'll work on adding the integration tests in my next change)
Fixes: 123987001
Change-Id: I2825ee6d4a759b585d28ee42c695699169f5bec8