New lint integration: Don't specify the variant
The integation in the Android Gradle Plugin only presents the lint
models for a single variant to lint anyway, so there's nothing to be
gained by explicitly specifying it.

This has the advantage of avoiding warnings of variants not found when
projects don't all have matching variant names, e.g. when there are java
library project dependencies.

Bug: 160392650
Test: Existing behaviour. No tests for the lack of warning.
Change-Id: I73ed0406fb6d02f828e5207780330a781f596ad1