Run MergeNativeLibsTask non-incrementally when changedInputs.size > 20
Previously, MergeNativeLibsTask could take several minutes to run when
changedInputs.size was large because the number of times it calls
FileUtils.isFileInDirectory was proportional to the number of changed
inputs.

With this change, we ignore the changed inputs and run the task from
scratch if changedInputs.size > 20. I chose 20 as the cutoff based on
experiments with a repro project.

This is a temporary workaround until a proper fix can be implemented.

Bug: 175337498
Test: tested performance manually
Change-Id: Ic8c716e32250c042d9910231bc5a2e5a3ee1d18a