Modify ExtractProfilerNativeDependenciesTask to extract .so files
Previously, ExtractProfilerNativeDependenciesTask output a directory
containing jars, but with this change, it outputs a directory with the
extracted .so files.

The reason for this change is to prepare for a subsequent change in
which the MergeNativeLibsTask logic will be simplified. After this
change, all inputs to MergeNativeLibsTask will be directories containing
.so files, so we won't need logic to handle jars in that task.

Bug: 175337498
Test: ExtractProfilerNativeDependenciesTaskTest, existing CustomClassTransformTest
Change-Id: I29a9ff46431b196c5e0587404031271c4aa87b98